(regression tests/cleanup) DynamicArrayTest: removed unnecessary boxing

+++ unit-tests/com/sun/javatest/util/DynamicArrayTest.java
@@ -91,7 +91,7 @@ public class DynamicArrayTest {
         Integer[] array = new Integer[count];
 
         for (int i = 0; i < count; i++) {
-            array[i] = Integer.valueOf(i);
+            array[i] = i;
         }
 
         return array;
@@ -104,15 +104,15 @@ public class DynamicArrayTest {
         Object[] back = null;
 
         // ---- remove beginning of even size array ----
-        back = DynamicArray.append(data1, Integer.valueOf(99));
+        back = DynamicArray.append(data1, 99);
 
         back = null;
         Object[] data2 = null;
-        back = DynamicArray.append(data2, Integer.valueOf(100));
+        back = DynamicArray.append(data2, 100);
 
         back = null;
         Object[] data3 = newIntArr(0);
-        back = DynamicArray.append(data3, Integer.valueOf(101));
+        back = DynamicArray.append(data3, 101);
 
         back = null;
         try {
@@ -475,7 +475,7 @@ public class DynamicArrayTest {
         // Integer and Float are not compatible
         try {
             Object[] data1 = newIntArr(2);
-            back = DynamicArray.append(data1, Float.valueOf(1.2f));
+            back = DynamicArray.append(data1, 1.2f);
             localResult = false;
             out.println("Stored inappropriate entry in array.");
         } catch (ArrayStoreException e) {
@@ -550,7 +550,7 @@ public class DynamicArrayTest {
         Integer[] back = null;
 
         // valid removal of an object
-        Integer[] data1 = {Integer.valueOf(311), Integer.valueOf(911)};
+        Integer[] data1 = {311, 911};
         back = DynamicArray.remove(data1, data1[0]);
 
         if (back.length != 1 || back[0] != data1[1]) {
@@ -591,8 +591,8 @@ public class DynamicArrayTest {
         Integer[] back = null;
 
         // insert object in the middle
-        Integer[] data1 = {Integer.valueOf(311), Integer.valueOf(911), Integer.valueOf(0)};
-        Integer newInt = Integer.valueOf(611);
+        Integer[] data1 = {311, 911, 0};
+        Integer newInt = 611;
         back = DynamicArray.insert(data1, newInt, 1);
 
         if (back.length != 4 || back[1] != newInt) {
@@ -604,7 +604,7 @@ public class DynamicArrayTest {
         back = null;
 
         // insert object at the beginning
-        Integer[] data2 = {Integer.valueOf(311), Integer.valueOf(911)};
+        Integer[] data2 = {311, 911};
         back = DynamicArray.insert(data2, newInt, 0);
 
         if (back.length != 3 || back[0] != newInt) {
@@ -616,7 +616,7 @@ public class DynamicArrayTest {
         back = null;
 
         // insert object at the end
-        Integer[] data3 = {Integer.valueOf(311), Integer.valueOf(911)};
+        Integer[] data3 = {311, 911};
         back = DynamicArray.insert(data3, newInt, 2);
 
         if (back.length != 3 || back[2] != newInt) {
@@ -647,7 +647,7 @@ public class DynamicArrayTest {
         }
 
         // check to see that is uses the supplied class if the orig. arr is null
-        arr = DynamicArray.append(arr, Integer.valueOf(1), intClass);
+        arr = DynamicArray.append(arr, 1, intClass);
 
         if (arr.getClass().getComponentType() != intClass) {
             out.println("New array type does not correspond to the requested type.");
@@ -655,7 +655,7 @@ public class DynamicArrayTest {
         }
 
         try {
-            arr = (Integer[]) DynamicArray.append(arr, Float.valueOf(1.0f), intClass);
+            arr = (Integer[]) DynamicArray.append(arr, 1.0f, intClass);
             // we are supposed to end up in the catch
             return false;
         } catch (java.lang.ArrayStoreException e) {
@@ -664,7 +664,7 @@ public class DynamicArrayTest {
 
         // check to see that it does not use the supplied class for a non-null orig. arr
         arr = new Integer[0];
-        arr = (Integer[]) DynamicArray.append(arr, Integer.valueOf(1), boolClass);
+        arr = (Integer[]) DynamicArray.append(arr, 1, boolClass);
 
         if (arr.getClass().getComponentType() != intClass) {
             out.println("Array type should not be the requested type.");
